Time: 520 ms (18.89%), Space: 90.4 MB (25.24%) - LeetHub

+++ 1248-count-number-of-nice-subarrays/1248-count-number-of-nice-subarrays.cpp
@@ -7,7 +7,7 @@ public:
        }   
       
         unordered_map<int,int> mpp;
-        int ans=0,prefix=0;
+        long long int ans=0,prefix=0;
         for(auto it:nums){
             prefix+=it;
             if(prefix==k) ans++;
